Readme and move multinew to multi
